added refactor, fix smell
